Change averageDifficulty type from Double to Integer
Updated averageDifficulty field across the codebase to use Integer instead of Double for consistency and simplicity. Modified relevant models, services, repositories, and controllers to reflect this change. Also updated Hibernate configuration to "create-drop" for testing changes.

+++ lorecodex/src/main/java/com/lorecodex/backend/controller/ChallengeController.java
@@ -271,7 +271,7 @@ public class ChallengeController {
     // Get average difficulty of a challenge
     @GetMapping("/{id}/average-difficulty")
     public ResponseEntity<Double> getAverageDifficulty(@PathVariable Long id) {
-        Double averageDifficulty = difficultyRatingService.getAverageDifficultyByChallengeId(id);
+        Integer averageDifficulty = difficultyRatingService.getAverageDifficultyByChallengeId(id);
         return ResponseEntity.ok(averageDifficulty != null ? averageDifficulty : 0.0);
     }
 }
